fix(logs): address Phase 7 review violations
- seqToken type changed from int to int64 (store, streams, events)
- UntagLogGroup nil-guard before deleting from tags map
- Register LogsCodec in codecMap so request routing works

+++ internal/provider/aws/cloudwatch/logs/events.go
@@ -104,7 +104,7 @@ func (p *Provider) PutLogEvents(_ context.Context, nr *model.NormalizedRequest) 
 	// Increment seq token
 	seqTokens := p.store.seqToken[groupName]
 	if seqTokens == nil {
-		seqTokens = make(map[string]int)
+		seqTokens = make(map[string]int64)
 		p.store.seqToken[groupName] = seqTokens
 	}
 	seqTokens[streamName]++
